% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_cadence_query.R
\name{bout_cadence_query}
\alias{bout_cadence_query}
\title{Bout cadence query}
\usage{
bout_cadence_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, bout_cadence_date, bout_cadence_value
}
\description{
Bout cadence query
}
\details{
Bout cadence is the average steps per minute when the step count >/= 60 steps a minute for at least 2 minutes.
The wearer will usually have many of these “bouts” throughout the day.
We take the average over the entire day to get the average bout cadence.
}
\examples{
\dontrun{
bout_dat <- bout_cadence_query()
}
}
