% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_code_query.R
\name{concept_code_query}
\alias{concept_code_query}
\title{Concept code query}
\usage{
concept_code_query(
  codes,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{codes}{a character vector or string containing medication names}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table containing the following columns:
person_id, condition_start_date, concept_name, condition_concept_id
}
\description{
Concept code query
}
\examples{
\dontrun{
concept_code_dat <- concept_code_query(c(441641,4014295))
}
}
