% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_query.R
\name{condition_query}
\alias{condition_query}
\title{Condition query}
\usage{
condition_query(
  concept_ids = NULL,
  source_values = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{concept_ids}{a numeric vector with condition concept ids}

\item{source_values}{a character vector with condition source values}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table containing the following columns:
person_id, condition_start_date, condition_concep_id, condition_source_value
}
\description{
Condition query
}
\examples{
\dontrun{
condition_dat <- condition_query(concept_ids = c(710706,705076), source_values = c("U09.9"))
}
}
