% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_query.R
\name{cpt_query}
\alias{cpt_query}
\title{CPT Query}
\usage{
cpt_query(cpt_codes, page_size = NULL)
}
\arguments{
\item{cpt_codes}{a character vector or character string of CPT codes}

\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
A data.table with the following columns:
person_id, cpt_code, entry_date
}
\description{
CPT Query
}
\examples{
\dontrun{
cpt_dat <- cpt_query(c("33510","33511"))
}
}
