% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_cause_value.R
\name{death_cause_query}
\alias{death_cause_query}
\title{Death cause query}
\usage{
death_cause_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, death_cause_entry_date, death_cause_value
}
\description{
Death cause query
}
\examples{
/dontrun{
death_dat <- death_cause_query()
}
}
