% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_query.R
\name{demographics_query}
\alias{demographics_query}
\title{Demographics query#'}
\usage{
demographics_query()
}
\value{
a data.table with the following columns:
person_id, date_of_birth, race, ethnicity, sex
}
\description{
Demographics query#'
}
\examples{
/dontrun{
dem_dat <- demographics_query()
}
}
