% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprivation_index_query.R
\name{deprivation_index_query}
\alias{deprivation_index_query}
\title{Deprivation index query}
\usage{
deprivation_index_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, deprivation_index_entry_date, deprivation_index_value
}
\description{
Deprivation index query
}
\examples{
/dontrun{
dep_idx <- deprivation_index_query()
}
}
