% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_query.R
\name{device_query}
\alias{device_query}
\title{Device query}
\usage{
device_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, device_type, device_model, data_record_date, last_sync_time, last_sync_date
anchored by last sync date
}
\description{
Device query
}
\examples{
/dontrun{
device_dat <- device_query()
}
}
