% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_big_data.R
\name{download_big_data}
\alias{download_big_data}
\title{Download big data}
\usage{
download_big_data(query, dest, rm_csv = TRUE)
}
\arguments{
\item{query}{SQL query string}

\item{dest}{name of the csv file}

\item{rm_csv}{logical indicating whether to delete the csv after downloading. Default is TRUE.}
}
\value{
If rm_csv is TRUE, it will return a data.table corresponding to the query only.
If rm_csv is FALSE, then in addition to the above, a csv file will also be saved to a folder called aou_reader in the workspace bucket.
}
\description{
Download big data
}
\examples{
\dontrun{
dataset <- Sys.getenv("WORKSPACE_CDR")
query <- stringr::str_glue("
       SELECT
           survey.person_id,
           survey.answer AS survey_response,
           CAST(survey.survey_datetime AS DATE) AS survey_date
       FROM
           `{dataset}.ds_survey` survey
       WHERE
           (
               question_concept_id IN (
                         1586198
               )
           )")
dest <- "survey.csv"
res <- download_big_data(query,dest)
}

}
