% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg_query.R
\name{ecg_query}
\alias{ecg_query}
\title{ECG query}
\usage{
ecg_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
"Q-T interval","P-R Interval","T wave axis","R wave axis","P wave axis","QRS duration",
"Q-T interval corrected","QRS complex Ventricles by EKG","P wave Atrium by EKG","Heart rate.beat-to-beat by EKG",
"Q-T interval corrected based on Bazett formula","QRS duration {Electrocardiograph lead}","QRS axis",
"R-R interval by EKG"
}
\description{
ECG query
}
\examples{
\dontrun{
ecg_dat <- ecg_query_query()
}
}
