% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_bp_query.R
\name{high_bp_query}
\alias{high_bp_query}
\title{High Blood Pressure Query}
\usage{
high_bp_query(page_size = NULL)
}
\arguments{
\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
A data.table with the following columns:
person_id, measurement_date
}
\description{
Provides the first date where BP was >=140/>=90
}
\examples{
\dontrun{
high_bp_dat <- high_bp_query()
}
}
