% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalization_query.R
\name{hospitalization_query}
\alias{hospitalization_query}
\title{Hospitalization query}
\usage{
hospitalization_query(
  codes = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{codes}{a character vector or character string of ICD codes. If NULL, all hospitalizations will be downloaded.}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns: person_id, hospitalization_entry_date, hospitalization_icd_code
}
\description{
Hospitalization query
}
\examples{
\dontrun{
dat <- hospitalization_query("I40")
}
}
