% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_min_heart_rate_query.R
\name{hourly_min_heart_rate_query}
\alias{hourly_min_heart_rate_query}
\title{Hourly minimum heart rate query}
\usage{
hourly_min_heart_rate_query(
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  cohort = NULL
)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{cohort}{a vector or list of integers indicating the person_id to be queried.}
}
\value{
a data.table with the following columns:
person_id, date, hour, min_heart_rate
}
\description{
Hourly minimum heart rate query
}
\details{
Returns the minimum heart rate at each day at each hour.
}
\examples{
\dontrun{
hr_df <- hourly_min_heart_rate_query()
}
}
