% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_steps_query.R
\name{hourly_steps_query}
\alias{hourly_steps_query}
\title{Hourly steps query}
\usage{
hourly_steps_query(
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  cohort = NULL
)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{cohort}{a vector or list of integers indicating the person_id to be queried.}
}
\value{
a data.table with the following columns:
person_id, date, hour, hourly_steps_count
}
\description{
Hourly steps query
}
\details{
Returns the number of steps at each day at each hour
}
\examples{
\dontrun{
hourly_steps_df <- hourly_steps_query()
}
}
