% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_query.R
\name{icd10_query}
\alias{icd10_query}
\title{ICD10 Query}
\usage{
icd10_query(dataset, icd10_codes, page_size = NULL)
}
\arguments{
\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}

\item{icd9_codes}{a character vector or character string containing ICD10 codes. String can contain wildcards using \% (e.g. "410.\%").}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
ICD10 Query
}
\examples{
\dontrun{
icd10_dat <- icd10_codes(c("I21","I21.\%"))
}
}
