% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_query.R
\name{icd9_query}
\alias{icd9_query}
\title{ICD9 Query}
\usage{
icd9_query(icd9_codes, page_size = NULL)
}
\arguments{
\item{icd9_codes}{a character vector or character string containing ICD9 codes. String can contain wildcards using \% (e.g. "410.\%").}

\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
ICD9 Query
}
\examples{
\dontrun{
icd9_dat <- icd9_codes(c("410","410.\%"))
}
}
