% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_procedure_query.R
\name{icd_procedure_query}
\alias{icd_procedure_query}
\title{ICD Procedure Query}
\usage{
icd_procedure_query(
  icd_codes = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{icd_codes}{a character vector or character string containing ICD procedure codes. String can contain wildcards using \%.}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, icd_procedure_date, icd_procedure_code
}
\description{
ICD Procedure Query
}
\examples{
\dontrun{
icd_proc_dat <- icd_procedure_query(c("0210\%","0211\%"))
}
}
