% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpatient_icd_query.R
\name{inpatient_icd_query}
\alias{inpatient_icd_query}
\title{Inpatient ICD query}
\usage{
inpatient_icd_query(
  codes,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{codes}{a character vector or string of ICD9/10 codes}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
Inpatient ICD query
}
\details{
Matches provided ICD9/10 code
AND
(
Can be any position on chart. Captured by condition_type_concept_id in condition_occurrence table
OR
Is an inpatient visit. Captured by the visit_concept_id in the visit_occurrence table
)
}
\examples{
\dontrun{
inpatient_dat <- inpatient_icd_query(c("I21","I21.\%"))
}
}
