% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpatient_visit_query.R
\name{inpatient_visit_query}
\alias{inpatient_visit_query}
\title{Inpatient visit query}
\usage{
inpatient_visit_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, inpatient_visit_date
}
\description{
Inpatient visit query
}
\details{
Provides the date of an inpatient visit.
}
\examples{
\dontrun{
inpatient_visit dat <- inpatient_visit_query()
}
}
