% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_concept_query.R
\name{lab_concept_query}
\alias{lab_concept_query}
\title{Lab concept query}
\usage{
lab_concept_query(
  lab_concepts,
  value_as_number = NULL,
  value_as_concept_id = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{value_as_number}{a numeric vector to subset labs by values as number}

\item{value_as_concept_id}{a numeric vector to subset labs by values as concept ids}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{labs}{a character vector or string containing the labs concepts to query}
}
\value{
a data.table containing the following columns:
person_id, measurement_date, value_as_number, value_as_concept
}
\description{
Lab concept query
}
\examples{
\dontrun{
lab_concept_dat <- lab_concept_query(c(586520,586523,586525))
}
}
