% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_query.R
\name{lab_query}
\alias{lab_query}
\title{Lab query}
\usage{
lab_query(labs, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{labs}{a character vector or string containing the labs to query}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table containing the following columns:
person_id, measurement_date, value_as_number
}
\description{
Lab query
}
\examples{
\dontrun{
lab_dat <- lab_query(c("Triglyceride [Mass/volume] in Serum or Plasma","Triglyceride [Mass/volume] in Blood"))
}
}
