% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_query_extended.R
\name{lab_query_extended}
\alias{lab_query_extended}
\title{Lab query with extended columns}
\usage{
lab_query_extended(
  labs,
  ext_cols = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{labs}{a character vector or string containing the labs to query}

\item{ext_cols}{extra columns to return on top of the date and value of the lab}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table containing the following columns:
person_id, measurement_date, value_as_number, ext_cols
}
\description{
Lab query with extended columns
}
\examples{
\dontrun{
lab_dat <- lab_query(c("Triglyceride [Mass/volume] in Serum or Plasma","Triglyceride [Mass/volume] in Blood"), ext_cols = c("unit_source_value"))
}
}
