% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_query.R
\name{med_query}
\alias{med_query}
\title{Medication query}
\usage{
med_query(meds, page_size = NULL)
}
\arguments{
\item{meds}{a character vector or string containing medication names}

\item{page_size}{= The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
a data.table containing the following columns:
person_id, drug_exposure_start_date
}
\description{
Medication query
}
\examples{
\dontrun{
med_dat <- med_query(c("roflumilast","daliresp"))
}
}
