% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpatient_icd_query.R
\name{outpatient_icd_query}
\alias{outpatient_icd_query}
\title{Outpatient ICD query}
\usage{
outpatient_icd_query(codes, page_size = NULL)
}
\arguments{
\item{codes}{a character vector or string of ICD9/10 codes}

\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
Outpatient ICD query
}
\details{
Matches provided ICD9/10 code
AND
(
Can be any position on chart. Captured by condition_type_concept_id in condition_occurrence table
OR
Is an outpatient visit. Captured by the visit_concept_id in the visit_occurrence table
)
}
\examples{
\dontrun{
outpatient_dat <- outpatient_icd_query(c("I21","I21.\%"))
}
}
