% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpatient_med_query.R
\name{outpatient_med_query}
\alias{outpatient_med_query}
\title{Outpatient Medication query}
\usage{
outpatient_med_query(
  meds,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{meds}{a character vector or string containing medication names}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table containing the following columns:
person_id, drug_exposure_start_date
}
\description{
Outpatient Medication query
}
\examples{
\dontrun{
med_dat <- outpatient_med_query(c("roflumilast","daliresp"))
}
}
