% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_consent_date_query.R
\name{primary_consent_date_query}
\alias{primary_consent_date_query}
\title{Primary consent date}
\usage{
primary_consent_date_query(
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
The date someone enrolled into the All of Us Research Program as a data contributing participant
a data.table with the following columns:
person_id, primary_consent_date
}
\description{
Primary consent date
}
\examples{
/dontrun{
consent_dat <- primary_consent_date_query()
}
}
