% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdann_query.R
\name{sdann_query}
\alias{sdann_query}
\title{SDANN query}
\usage{
sdann_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, sdann_date, sdann_value, sdann_total_valid_interval
}
\description{
SDANN query
}
\details{
Heart rate minute level data is extracted and five-minute averages are calculated for consecutive five-minute intervals.
The average HR is used to calculate the average RR duration of each five-minute interval:
Average RR = 6000 / mean(HR)
Subsequently, the standard deviation of all the five-minute RR intervals is calculated, yielding the SDANN value (in ms).
}
\examples{
/dontrun{
sdann_dat <- sdann_query()
}
}
