% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_level_query.R
\name{sleep_level_query}
\alias{sleep_level_query}
\title{Sleep level query}
\usage{
sleep_level_query(
  sleep_terms,
  time = "first",
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{sleep_terms}{a SQL WHERE statement for filtering sleep levels. If no statement is provided, then no filtering is performed.}

\item{time}{which time should be pulled? Can be \code{first}, \code{last}, or \code{all}. Note, \code{all} will return all sleep levels, which can be quite large. Default is \code{first}.}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, sleep_level_date, sleep_level_datetime, sleep_level_duration, sleep_level_is_main_sleep
}
\description{
Sleep level query
}
\examples{
/dontrun{
sleep_dat <- sleep_query("is_main_sleep = TRUE AND level = 'deep'") #first deep/main sleep datetimes
}
}
