% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_query.R
\name{sleep_query}
\alias{sleep_query}
\title{Sleep data from Fitbit query}
\usage{
sleep_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, sleep_date, is_main_sleep, minute_in_bed, minute_asleep,
minute_after_wakeup, minute_awake, minute_restless,
minute_deep, minute_light, minute_rem, minute_wake
}
\description{
Sleep data from Fitbit query
}
\examples{
/dontrun{
sleep_dat <- sleep_query()
}
}
