% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snomed_query.R
\name{snomed_query}
\alias{snomed_query}
\title{snomed Query}
\usage{
snomed_query(
  snomed_codes = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{snomed_codes}{a character vector or character string containing snomed codes. String can contain wildcards using \% (e.g. "410.\%").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
snomed Query
}
\examples{
\dontrun{
snomed_dat <- snomed_codes(c("410","410.\%"))
}
}
