% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_of_residence_query.R
\name{state_of_residence_query}
\alias{state_of_residence_query}
\title{State of residence query}
\usage{
state_of_residence_query()
}
\value{
a data.table with the following columns:
person_id, state_of_residence_value
}
\description{
State of residence query
}
\examples{
/dontrun{
sor_dat <- state_of_residence_query()
}
}
