% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_count_5_minute_level_query.R
\name{step_count_average_5_minute_interval_query}
\alias{step_count_average_5_minute_interval_query}
\title{Average step count over all 5-minute-intervals query}
\usage{
step_count_average_5_minute_interval_query()
}
\value{
a data.table with the following columns:
person_id, minute_interval, average_step_count, p_has_steps
p is the probability of the interval having > 0 steps.
}
\description{
Average step count over all 5-minute-intervals query
}
\details{
Average step count over all 5-minute intervals are computed for each person_id.
}
\examples{
/dontrun{
step_dat <- step_count_average_5_minute_interval_query()
}
}
