% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_query.R
\name{survey_query}
\alias{survey_query}
\title{Survey query}
\usage{
survey_query(survey_codes, page_size = NULL)
}
\arguments{
\item{survey_codes}{a character vector or string of survey codes}

\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
a data.table with the following columns:
person_id, survey_response, survey_date
}
\description{
Survey query
}
\examples{
\dontrun{
survey_dat <- survey_query("1585860")
}
}
