% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_steps_query.R
\name{top_steps_query}
\alias{top_steps_query}
\title{Top steps query}
\usage{
top_steps_query(
  top_number,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{top_number}{Top x-minute step count, where top_number = x.}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
Top steps query
}
\details{
Top x-minute step count for each day.
}
\examples{
\dontrun{
top_steps_dat <- top_steps(1)
}
}
