% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_study_query.R
\name{wear_study_query}
\alias{wear_study_query}
\title{Wear study query}
\usage{
wear_study_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, wear_study_consent, wear_study_consent_start_date, wear_study_consent_end_date
}
\description{
Wear study query
}
\examples{
/dontrun{
wear_study_dat <- wear_study_query()
}
}
