% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_time_query.R
\name{wear_time_query}
\alias{wear_time_query}
\title{Wear time query}
\usage{
wear_time_query(anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
Wear time query
}
\details{
An hour of wear time is defined when step count is > 0 for a given hour of the day,
}
\examples{
\dontrun{
wear_time_date <- wear_time_query()
}
}
