% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_query.R
\name{wgs_query}
\alias{wgs_query}
\title{Whole genome sequencing cohort query}
\usage{
wgs_query()
}
\value{
a data.table with the following columns:
person_id
}
\description{
Whole genome sequencing cohort query
}
\examples{
\dontrun{
wgs_dat <- wgs_query()
}
}
