% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_big_data_robust.R
\name{download_big_data_robust}
\alias{download_big_data_robust}
\title{Download big data via BigQuery extract to GCS (robust for huge results)}
\usage{
download_big_data_robust(query, dest, rm_csv = TRUE)
}
\description{
Download big data via BigQuery extract to GCS (robust for huge results)
}
