% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_query.R
\name{icd10_query}
\alias{icd10_query}
\title{ICD10 Query}
\usage{
icd10_query(icd10_codes, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{icd10_codes}{a character vector or character string containing ICD10 codes. String can contain wildcards using \% (e.g. "410.\%").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
ICD10 Query
}
\examples{
\dontrun{
icd10_dat <- icd10_codes(c("I21","I21.\%"))
}
}
