% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_query.R
\name{icd9_query}
\alias{icd9_query}
\title{ICD9 Query}
\usage{
icd9_query(icd9_codes, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{icd9_codes}{a character vector or character string containing ICD9 codes. String can contain wildcards using \% (e.g. "410.\%").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, condition_start_date, condition_source_value
}
\description{
ICD9 Query
}
\examples{
\dontrun{
icd9_dat <- icd9_codes(c("410","410.\%"))
}
}
