% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_query.R
\name{lab_query}
\alias{lab_query}
\title{Lab query}
\usage{
lab_query(labs, page_size = NULL)
}
\arguments{
\item{labs}{a character vector or string containing the labs to query}

\item{page_size}{The number of rows requested per chunk. It is recommended to leave this unspecified (see bq_table_download {bigrquery}).}
}
\value{
a data.table containing the following columns:
person_id, measurement_date, value_as_number
}
\description{
Lab query
}
\examples{
\dontrun{
lab_dat <- lab_query(c("Triglyceride [Mass/volume] in Serum or Plasma","Triglyceride [Mass/volume] in Blood"))
}
}
