% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_query.R
\name{survey_query}
\alias{survey_query}
\title{Survey query}
\usage{
survey_query(
  survey_codes,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{survey_codes}{a character vector or string of survey codes}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
a data.table with the following columns:
person_id, survey_response, survey_date
}
\description{
Survey query
}
\examples{
\dontrun{
survey_dat <- survey_query("1585860")
}
}
